requirements.txt
Flask==2.2.5
python-dotenv==1.0.0
Pillow==9.4.0
fpdf==1.7.2
markdown==3.4.3
google-generativeai==0.1.0
requests==2.28.1